+++ android/app/src/main/kotlin/deckers/thibault/aves/utils/MimeTypes.kt
@@ -1,7 +1,7 @@
 package deckers.thibault.aves.utils
 
 import android.webkit.MimeTypeMap
-import deckers.thibault.aves.decoder.MultiPageImage
+import deckers.thibault.aves.glide.MultiPageImage
 import androidx.exifinterface.media.ExifInterfaceFork as ExifInterface
 
 object MimeTypes {
@@ -84,11 +84,9 @@ object MimeTypes {
         else -> false
     }
 
-    // as of Flutter v3.16.4, with additional custom handling for SVG in Dart,
-    // while handling still PNG and JPEG on Android for color space and config conversion
-    fun canDecodeWithFlutter(mimeType: String, isAnimated: Boolean) = when (mimeType) {
-        GIF, WEBP, BMP, WBMP, ICO, SVG -> true
-        JPEG, PNG -> isAnimated
+    // as of Flutter v3.16.4, with additional custom handling for SVG in Dart
+    fun handleEncodedBytesInFlutter(mimeType: String) = when (mimeType) {
+        JPEG, PNG, GIF, WEBP, BMP, WBMP, ICO, SVG -> true
         else -> false
     }