+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+## <a id="v1.13.8"></a>[v1.13.8] - 2025-09-08
+
 ### Added
 
 - Collection: support for Motorola burst pattern
@@ -12,12 +14,15 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - manual rotation to landscape switches to current sensor orientation, then is fixed to it
+- improved image streaming and pixel format selection
 
 ### Fixed
 
 - receiving `geo:` URI in some cases
 - visibility of hidden new albums
 - customizing cover or hiding tag groups
+
+## <a id="v1.13.7"></a>[v1.13.7] - 2025-09-07 [YANKED]
 
 ## <a id="v1.13.6"></a>[v1.13.6] - 2025-07-28